[RFC][pxa] Draft of layout reorder pass
Basic flow and concept of a pass performing optimization of memory by
changing layout.
Main aim of this change is to show general concept, some details of
the implementation may change.
As of now the plan is to perform this optimization right after
subgroup/tile pass. In future it probably should be tightly integrated
with those passes.

Future roadmap, not included in this draft, would encompass:
- adding support for memory with multiple users
- taking into account parallelization when sorting dimensions
- adding heuristic for cache misses and determining if separate reorder
  is optimal when cannot fuse into producer
- integrating with tiling pass - estimate tiling of loops together with
  tiling memory
- taking into account constness of data (ie. weights) when selecting
  which memory to reorder